Develop Soft Delete function at Post App's Post/Comment Entity

See merge request geuun/geun-sns!31

+++ src/main/java/com/dev/geunsns/apps/post/data/dto/postlike/response/PostLikeCountResponse.java
@@ -9,10 +9,10 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class PostLikeCountResponse {
     private Long postId;
-    private Long likeCount;
+    private Integer likeCount;
 
     @Builder
-    public PostLikeCountResponse(Long postId, Long likeCount) {
+    public PostLikeCountResponse(Long postId, Integer likeCount) {
         this.postId = postId;
         this.likeCount = likeCount;
     }
